+++ uemf.h
@@ -840,7 +840,7 @@ extern int		cronFree(cron_t *cp);
 #define SOCKET_BUFSIZ			1024	/* Underlying buffer size */
 #endif /* LITTLEFOOT */
 
-typedef void 	(*socketHandler_t)(int sid, int mask, int data);
+typedef void 	(*socketHandler_t)(int sid, int mask, long data);
 typedef int		(*socketAccept_t)(int sid, char *ipaddr, int port, 
 					int listenSid);
 typedef struct {
@@ -850,7 +850,7 @@ typedef struct {
 	ringq_t			lineBuf;				/* Line ring queue */
 	socketAccept_t	accept;					/* Accept handler */
 	socketHandler_t	handler;				/* User I/O handler */
-	int				handler_data;			/* User handler data */
+	long			handler_data;			/* User handler data */
 	int				handlerMask;			/* Handler events of interest */
 	int				sid;					/* Index into socket[] */
 	int				port;					/* Port to listen on */
@@ -1009,7 +1009,7 @@ extern int		scriptEval(int engine, char_t *cmd, char_t **rslt, int chan);
 extern void		socketClose();
 extern void		socketCloseConnection(int sid);
 extern void		socketCreateHandler(int sid, int mask, socketHandler_t 
-					handler, int arg);
+					handler, long arg);
 extern void		socketDeleteHandler(int sid);
 extern int		socketEof(int sid);
 extern int 		socketCanWrite(int sid);
